% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{load_functions}
\alias{load_functions}
\title{Source R Functions from Directory}
\usage{
load_functions()
}
\value{
Invisible NULL. The function sources the R files for their side effects.
}
\description{
Sources R files listed in `_project.yml` and from fallback folders.
}
\examples{
\dontrun{
# Source all R files from default directory
load_functions()
}
}
