% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_fun}
\alias{project_fun}
\title{Source R Functions from Directory}
\usage{
project_fun(
  folders = c("Rcode/function", "Rcode/00_function", "script/function",
    "script/00_function", "../000-CommonFiles/script/00_function")
)
}
\arguments{
\item{folders}{Character vector. Paths to the directories containing R files relative to project root. 
Default is c("Rcode/function",  "Rcode/00_function", "script/function", "script/00_function",  "../000-CommonFiles/script/00_function").}
}
\value{
Invisible NULL. The function sources the R files for their side effects.
}
\description{
This function sources all R files from a specified directory and its subdirectories into global environment.
}
\examples{
\dontrun{
# Source all R files from default directory
project_fun()
}
}
