% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks_mermaid.R
\name{read_asset_file}
\alias{read_asset_file}
\title{Read an asset file from the package inst/assets folder}
\usage{
read_asset_file(filename)
}
\arguments{
\item{filename}{The name of the asset file (e.g., "mermaid-zoom.css").}
}
\value{
A character string containing the file contents.
}
\description{
This helper function loads the contents of a file from the package's inst/assets directory.
It is used to include external CSS or JS in knitr output.
}
\examples{
# Read CSS asset
read_asset_file("mermaid-zoom.css")
}
