% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{load_functions}
\alias{load_functions}
\title{Source R Functions from Directory}
\usage{
load_functions(
  folders = c("Rcode/function", "Rcode/00_function", "script/function",
    "script/00_function")
)
}
\arguments{
\item{folders}{Character vector. Paths to the directories containing R files relative to project root. 
Default is c("Rcode/function",  "Rcode/00_function", "script/function", "script/00_function").}
}
\value{
Invisible NULL. The function sources the R files for their side effects.
}
\description{
Sources R files listed in `_project.yml` and from fallback folders.
}
\examples{
\dontrun{
# Source all R files from default directory
load_functions()
}
}
